+++ src/System.ValueTuple/src/System.ValueTuple.csproj
@@ -17,7 +17,7 @@
     <RootNamespace>System</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="System\Runtime\CompilerServices\TupleElementNamesAttribute.cs" />
+    <Compile Include="System\Runtime\CompilerServices\TupleItemNamesAttribute.cs" />
     <Compile Include="System\ValueTuple\TupleExtensions.cs" />
     <Compile Include="System\ValueTuple\ValueTuple.cs" />
   </ItemGroup>
