+++ src/System.ValueTuple/tests/System.ValueTuple.Tests.csproj
@@ -10,7 +10,7 @@
     <NugetTargetMoniker>.NETStandard,Version=v1.3</NugetTargetMoniker>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="TupleElementNamesAttribute\UnitTests.cs" />
+    <Compile Include="TupleItemNamesAttribute\UnitTests.cs" />
     <Compile Include="ValueTuple\ExtensionsTests.cs" />
     <Compile Include="ValueTuple\UnitTests.cs" />
   </ItemGroup>
